+++ packages/plugin-vue/src/main.ts
@@ -162,7 +162,7 @@ export async function transformMain(
   if (options.sourceMap) {
     if (scriptMap && templateMap) {
       // if the template is inlined into the main module (indicated by the presence
-      // of templateMap, we need to concatenate the two source maps.
+      // of templateMap), we need to concatenate the two source maps.
 
       const gen = fromMap(
         // version property of result.map is declared as string
@@ -221,7 +221,11 @@ export async function transformMain(
     const { code, map } = await transformWithEsbuild(
       resolvedCode,
       filename,
-      { loader: 'ts', sourcemap: options.sourceMap },
+      {
+        loader: 'ts',
+        target: 'esnext',
+        sourcemap: options.sourceMap
+      },
       resolvedMap
     )
     resolvedCode = code