+++ packages/vite/types/anymatch.d.ts
@@ -1,5 +1,1 @@
-export type AnymatchFn = (testString: string) => boolean
-export type AnymatchPattern = string | RegExp | AnymatchFn
-type AnymatchMatcher = AnymatchPattern | AnymatchPattern[]
-
-export { AnymatchMatcher as Matcher }
+export type { AnymatchFn, AnymatchPattern, Matcher } from '../dist/node'