+++ playground/css/main.js
@@ -3,6 +3,9 @@ import './minify.css'
 import css from './imported.css'
 text('.imported-css', css)
 
+import sugarss from './sugarss.sss'
+text('.imported-sugarss', sugarss)
+
 import sass from './sass.scss'
 text('.imported-sass', sass)
 
@@ -43,6 +46,8 @@ text('.modules-inline', inlineMod)
 
 import charset from './charset.css'
 text('.charset-css', charset)
+
+import './layered/index.css'
 
 import './dep.css'
 import './glob-dep.css'
@@ -107,3 +112,5 @@ document
   .classList.add(aliasModule.aliasedModule)
 
 import './unsupported.css'
+
+import './async/index'