+++ playground/vue-lib/src-lib-css/index.ts
@@ -1,0 +1,3 @@
+import './index.css'
+
+export function setup() {}